[optimize] rename non_cuda_equivalents to python_ops_fallback
