+++ docs/design/ARCHITECTURE_MULTI_HARDWARE.md
@@ -69,7 +69,7 @@
 │ │ • Layerwise     │  │ • LayerwiseXPU  │  │                 │   │
 │ │ • Buffer        │  │                 │  │ torch.hpu.*     │   │
 │ │ • SGLang        │  │ torch.xpu.*     │  │                 │   │
-│ │                 │  │ non_cuda_equiv  │  │                 │   │
+│ │                 │  │ python_ops_fb   │  │                 │   │
 │ │ torch.cuda.*    │  │                 │  │                 │   │
 │ │ c_ops + cupy    │  │                 │  │                 │   │
 │ └─────────────────┘  └─────────────────┘  └─────────────────┘   │
@@ -132,5 +132,5 @@ which is wrong for that backend's actual KV cache layout.
 1. Add detection branch in `__init__.py` `_detect_device()`
 2. Create `gpu_connector/xxx_connectors.py`, implement `GPUConnectorInterface`
 3. Add routing branch in `gpu_connector/__init__.py`
-4. Add kernels in `c_ops/` or fallback in `non_cuda_equivalents.py`
+4. Add kernels in `c_ops/` or fallback in `python_ops_fallback.py`
 5. No changes needed in middle layer code
